Converted some files to jsx, refactored navbar and added functionality to redirect after user logs in
